CIP-3407 pull changes in

+++ src/DelayModelStore.h
@@ -9,10 +9,10 @@ struct FoPoly
 	double ho_poly_start_time_ms;
 	double start_time_ms;
 	double stop_time_ms;
-	double poly[2];     // Units: ns/s^(num_coeffs - index - 1)
+	long double poly[2];     // Units: ns/s^(num_coeffs - index - 1)
 
-    inline double delay_const() { return poly[1]; }
-    inline double delay_linear() { return poly[0]; }
+    inline long double delay_const() { return poly[1]; }
+    inline long double delay_linear() { return poly[0]; }
 };
